Update build_errors.md
Corrected a typo (uninstal with one l) but more importantly the code to install homebrew on this Build Errors page points to an outdated version of Homebrew. The correct install command is the one in Step 7 which I have copied here. What I Cannot help with is that the uninstall command probably needs to be updated as well. I am far from a programmer but what I understood from the response (in terminal) to submitting the outdated code indicates (in comments) that the makers of Homebrew moved from the "RUBY" to "BASH", so the fact the uninstall command mentions "ruby" makes me think it is also outdated. The uninstall command in Step 7 also has a "ruby" reference so it may need to be changed there as well.