Update step7.md
Prior to this I submitted a pull request for the build errors page. This is related to my suggestion there. I am far from a programmer but what I understood from the response (in terminal) to submitting the outdated code to install Homebrew, the makers indicate (in comments) that they had moved from "RUBY" to "BASH". Here in Step 7 you correctly state the newer "BASH" command to install Homebrew but in Build Errors the older "RUBY" command is listed (that's what that pull request addresses). Here, however, the fact the uninstall command mentions "ruby" makes me think it is also outdated. I don't know enough to know what the correct uninstall command would be (or perhaps for uninstalling it is fine to refer to RUBY--I don't know), but I am flagging this for the attention of one who would know. Thanks!